requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.1.4
djangorestframework==3.14.0
future==0.18.2
pytz==2022.7
serializers==0.2.4
sqlparse==0.4.3
